requirements.txt
pycryptodome>=3.17
pytest>=7.0
